+++ rmq_metadata.py
@@ -277,10 +277,10 @@ def validate_create_settings(cfg):
         settings.
 
     Arguments:
-        (input) cfg -> Configuration module name.
-        (output) cfg -> Configuration module handler.
-        (output) status_flag -> True|False - successfully validation/creation.
-        (output) err_msg -> Error message from checks.
+        (input) cfg -> Configuration module name
+        (output) cfg -> Configuration module handler
+        (output) status_flag -> True|False - successfully validation/creation
+        (output) err_msg -> Error message from checks
 
     """
 
@@ -359,11 +359,11 @@ def validate_files(cfg, status_flag, err_msg):
     Description:  Validates the file entries in the configuration file.
 
     Arguments:
-        (input) cfg -> Configuration settings module for the program.
-        (input) status_flag -> True|False - successfully validation.
-        (input) err_msg -> Error message from checks.
-        (output) status_flag -> True|False - successfully validation.
-        (output) err_msg -> Error message from checks.
+        (input) cfg -> Configuration settings module for the program
+        (input) status_flag -> True|False - successfully validation
+        (input) err_msg -> Error message from checks
+        (output) status_flag -> True|False - successfully validation
+        (output) err_msg -> Error message from checks
 
     """
 
@@ -405,12 +405,12 @@ def non_proc_msg(rmq, log, cfg, data, subj, r_key):
     Description:  Process non-processed messages.
 
     Arguments:
-        (input) rmq -> RabbitMQ class instance.
-        (input) log -> Log class instance.
-        (input) cfg -> Configuration settings module for the program.
-        (input) data -> Body of message that was not processed.
-        (input) subj -> Email subject line.
-        (input) r_key -> Routing key for message.
+        (input) rmq -> RabbitMQ class instance
+        (input) log -> Log class instance
+        (input) cfg -> Configuration settings module for the program
+        (input) data -> Body of message that was not processed
+        (input) subj -> Email subject line
+        (input) r_key -> Routing key for message
 
     """
 
@@ -456,11 +456,11 @@ def process_msg(rmq, log, cfg, method, body):
     Description:  Process message from RabbitMQ queue.
 
     Arguments:
-        (input) rmq -> RabbitMQ class instance.
-        (input) log -> Log class instance.
-        (input) cfg -> Configuration settings module for the program.
-        (input) method -> Delivery properties.
-        (input) body -> Message body.
+        (input) rmq -> RabbitMQ class instance
+        (input) log -> Log class instance
+        (input) cfg -> Configuration settings module for the program
+        (input) method -> Delivery properties
+        (input) body -> Message body
 
     """
 
@@ -504,12 +504,12 @@ def convert_data(rmq, log, cfg, queue, body, r_key):
     Description:  Pre-processing of message and decode the message.
 
     Arguments:
-        (input) rmq -> RabbitMQ class instance.
-        (input) log -> Log class instance.
-        (input) cfg -> Configuration settings module for the program.
-        (input) queue -> RabbitMQ queue.
-        (input) body -> Message body.
-        (input) r_key -> Routing key.
+        (input) rmq -> RabbitMQ class instance
+        (input) log -> Log class instance
+        (input) cfg -> Configuration settings module for the program
+        (input) queue -> RabbitMQ queue
+        (input) body -> Message body
+        (input) r_key -> Routing key
 
     """
 
@@ -569,10 +569,10 @@ def read_pdf(filename, log):
     Description:  Extract text from a PDF file using the PyPDF2 module.
 
     Arguments:
-        (input) filename -> PDF file name.
-        (input) log -> Log class instance.
-        (output) status -> True|False - successfully extraction of data.
-        (output) text -> Raw text.
+        (input) filename -> PDF file name
+        (input) log -> Log class instance
+        (output) status -> True|False - successfully extraction of data
+        (output) text -> Raw text
 
     """
 
@@ -606,9 +606,9 @@ def find_tokens(tokenized_text, cfg):
         tokens.
 
     Arguments:
-        (input) tokenized_text -> List of tokens.
-        (input) cfg -> Configuration settings module for the program.
-        (output) categorized_text -> List of categorized tokens.
+        (input) tokenized_text -> List of tokens
+        (input) cfg -> Configuration settings module for the program
+        (output) categorized_text -> List of categorized tokens
 
     """
 
@@ -627,9 +627,9 @@ def summarize_data(categorized_text, token_types):
         a single unique list.
 
     Arguments:
-        (input) categorized_text -> List of categorized tokens.
-        (input) token_types -> List of token types to be accepted.
-        (output) data_list -> List of summarized categorized tokens.
+        (input) categorized_text -> List of categorized tokens
+        (input) token_types -> List of token types to be accepted
+        (output) data_list -> List of summarized categorized tokens
 
     """
 
@@ -658,14 +658,14 @@ def sort_data(item, current_type, data_list, tmp_data, token_types):
         ignores the "O" (OTHER) token type.
 
     Arguments:
-        (input) item -> Single set token.
-        (input) current_type -> Current token type.
-        (input) data_list -> List of summarized categorized tokens.
-        (input) tmp_data -> List of current series of token data.
-        (input) token_types -> List of token types.
-        (output) current_type -> Current token type.
-        (output) data_list -> List of summarized categorized tokens.
-        (output) tmp_data -> List of current series of token data.
+        (input) item -> Single set token
+        (input) current_type -> Current token type
+        (input) data_list -> List of summarized categorized tokens
+        (input) tmp_data -> List of current series of token data
+        (input) token_types -> List of token types
+        (output) current_type -> Current token type
+        (output) data_list -> List of summarized categorized tokens
+        (output) tmp_data -> List of current series of token data
 
     """
 
@@ -703,9 +703,9 @@ def merge_data(data_list, tmp_data):
         and adds the token type and string as set to a list.
 
     Arguments:
-        (input) data_list -> List of summarized categorized tokens.
-        (input) tmp_data -> List of current series of token data.
-        (output) data_list -> List of summarized categorized tokens.
+        (input) data_list -> List of summarized categorized tokens
+        (input) tmp_data -> List of current series of token data
+        (output) data_list -> List of summarized categorized tokens
 
     """
 
@@ -731,11 +731,11 @@ def get_pypdf2_data(f_name, cfg, log):
         file extracted using PyPDF2 module.
 
     Arguments:
-        (input) f_name -> PDF file name.
-        (input) cfg -> Configuration settings module for the program.
-        (input) log -> Log class instance.
-        (output) status -> True|False - successfully extraction of data.
-        (output) final_data -> List of categorized tokens from PDF file.
+        (input) f_name -> PDF file name
+        (input) cfg -> Configuration settings module for the program
+        (input) log -> Log class instance
+        (output) status -> True|False - successfully extraction of data
+        (output) final_data -> List of categorized tokens from PDF file
 
     """
 
@@ -768,9 +768,9 @@ def create_metadata(metadata, data):
         based on the data set in the list.
 
     Arguments:
-        (input) metadata -> Dictionary of meta-data.
-        (input) data -> List of data sets.
-        (output) metadata -> Dictionary of meta-data.
+        (input) metadata -> Dictionary of meta-data
+        (input) data -> List of data sets
+        (output) metadata -> Dictionary of meta-data
 
     """
 
@@ -796,11 +796,11 @@ def extract_pdf(f_name, log, char_encoding=None):
     Description:  Extract text from PDF using textract module.
 
     Arguments:
-        (input) f_name -> PDF file name.
-        (input) log -> Log class instance.
-        (input) char_encoding -> Character encoding code.
-        (output) status -> True|False - successfully extraction of data.
-        (output) text -> Raw text.
+        (input) f_name -> PDF file name
+        (input) log -> Log class instance
+        (input) char_encoding -> Character encoding code
+        (output) status -> True|False - successfully extraction of data
+        (output) text -> Raw text
 
     """
 
@@ -848,11 +848,11 @@ def get_textract_data(f_name, cfg, log):
     Description:  Process data using the textract module.
 
     Arguments:
-        (input) f_name -> PDF file name.
-        (input) cfg -> Configuration settings module for the program.
-        (input) log -> Log class instance.
-        (output) status -> True|False - successfully extraction of data.
-        (output) final_data -> List of categorized tokens from PDF file.
+        (input) f_name -> PDF file name
+        (input) cfg -> Configuration settings module for the program
+        (input) log -> Log class instance
+        (output) status -> True|False - successfully extraction of data
+        (output) final_data -> List of categorized tokens from PDF file
 
     """
 
@@ -919,10 +919,10 @@ def pdf_to_string(f_name, log):
     Description:  Extract text from PDF using pdfminer module.
 
     Arguments:
-        (input) f_name -> PDF file name.
-        (input) log -> Log class instance.
-        (output) status -> True|False - successfully extraction of data.
-        (output) text -> Raw text.
+        (input) f_name -> PDF file name
+        (input) log -> Log class instance
+        (output) status -> True|False - successfully extraction of data
+        (output) text -> Raw text
 
     """
 
@@ -960,11 +960,11 @@ def get_pdfminer_data(f_name, cfg, log):
     Description:  Process data using the pdfminer module.
 
     Arguments:
-        (input) f_name -> PDF file name.
-        (input) cfg -> Configuration settings module for the program.
-        (input) log -> Log class instance.
-        (output) status -> True|False - successfully extraction of data.
-        (output) final_data -> List of categorized tokens from PDF file.
+        (input) f_name -> PDF file name
+        (input) cfg -> Configuration settings module for the program
+        (input) log -> Log class instance
+        (output) status -> True|False - successfully extraction of data
+        (output) final_data -> List of categorized tokens from PDF file
 
     """
 
@@ -995,11 +995,11 @@ def process_message(queue, cfg, f_name, log):
     Description:  Extract metadata from message.
 
     Arguments:
-        (input) queue -> RabbitMQ queue.
-        (input) cfg -> Configuration settings module for the program.
-        (input) f_name -> PDF file name.
-        (input) log -> Log class instance.
-        (output) status -> True|False - successfully extraction of data.
+        (input) queue -> RabbitMQ queue
+        (input) cfg -> Configuration settings module for the program
+        (input) f_name -> PDF file name
+        (input) log -> Log class instance
+        (output) status -> True|False - successfully extraction of data
 
     """
 
@@ -1063,8 +1063,8 @@ def monitor_queue(cfg, log):
     Description:  Monitor RabbitMQ queue for messages.
 
     Arguments:
-        (input) cfg -> Configuration settings module for the program.
-        (input) log -> Log class instance.
+        (input) cfg -> Configuration settings module for the program
+        (input) log -> Log class instance
 
     """
 
@@ -1075,10 +1075,10 @@ def monitor_queue(cfg, log):
         Description:  Process message from RabbitMQ.
 
         Arguments:
-            (input) channel -> Channel properties.
-            (input) method -> Delivery properties.
-            (input) properties -> Properties of the message.
-            (input) body -> Message body.
+            (input) channel -> Channel properties
+            (input) method -> Delivery properties
+            (input) properties -> Properties of the message
+            (input) body -> Message body
 
         """
 
@@ -1141,7 +1141,7 @@ def monitor_queue(cfg, log):
         log.log_err("Failed to connnect to RabbuitMQ -> Msg: %s" % (err_msg))
 
 
-def run_program(args_array, func_dict, **kwargs):
+def run_program(args, func_dict, **kwargs):
 
     """Function:  run_program
 
@@ -1149,16 +1149,16 @@ def run_program(args_array, func_dict, **kwargs):
         Set a program lock to prevent other instantiations from running.
 
     Arguments:
-        (input) args_array -> Dict of command line options and values.
-        (input) func_dict -> Dict of function calls and associated options.
+        (input) args -> ArgParser class instance.
+        (input) func_dict -> Dict of function calls and associated options
 
     """
 
     cmdline = gen_libs.get_inst(sys)
-    args_array = dict(args_array)
+#    args_array = dict(args_array)
     func_dict = dict(func_dict)
-    cfg = gen_libs.load_module(args_array["-c"], args_array["-d"])
-    cfg.mongo = gen_libs.load_module(cfg.mongo_cfg, args_array["-d"])
+    cfg = gen_libs.load_module(args.get_val("-c"), args.get_val("-d"))
+    cfg.mongo = gen_libs.load_module(cfg.mongo_cfg, args.get_val("-d"))
     cfg, status_flag, err_msg = validate_create_settings(cfg)
 
     if status_flag:
@@ -1179,11 +1179,11 @@ def run_program(args_array, func_dict, **kwargs):
         log.log_info("%s" % (str_val))
 
         try:
-            flavor_id = args_array.get("-y", cfg.exchange_name)
+            flavor_id = args.get_val("-y", def_val=cfg.exchange_name)
             prog_lock = gen_class.ProgramLock(cmdline.argv, flavor_id)
 
-            # Intersect args_array & func_dict to find which functions to call.
-            for opt in set(args_array.keys()) & set(func_dict.keys()):
+            # Intersect args.args_array & func_dict to determine function call
+            for opt in set(args.get_args_keys()) & set(func_dict.keys()):
                 func_dict[opt](cfg, log, **kwargs)
 
             del prog_lock