+++ CHANGELOG.md
@@ -14,7 +14,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - validate_create_settings:  Incorrect setting of tmp_dir if not absolute path setting.
 
 ### Changed
-- main: Replaced the use of arg_parser (args_array) with gen_class.ArgParser class (args).
+- main, run_program: Replaced the use of arg_parser (args_array) with gen_class.ArgParser class (args).
 - Renamed \_validate_files to validate_files.
 - Renamed \_sort_data to sort_data.
 - Renamed \_process_queue to process_message.